Handle failed cutechess runs

+++ tasks/games.py
@@ -11,10 +11,10 @@ def run_games(run_id, run_chunk):
   rundb = RunDb()
   run = rundb.get_run(run_id)
 
-  # Create temporary directory, and copy everything in.  This is to allow multiple
-  # tasks to run at once
   use_temp_dir = False
   if use_temp_dir:
+    # Create temporary directory, and copy everything in.  This is to allow multiple
+    # tasks to run at once
     working_dir = tempfile.mkdtemp()
     stockfish_dir = os.path.join(working_dir, 'stockfish')
     testing_dir = os.path.join(working_dir, 'testing')
@@ -42,7 +42,7 @@ def run_games(run_id, run_chunk):
   sh.cd(testing_dir)
   sh.rm('-f', 'results.pgn')
 
-  state = None
+  state = {'wins':0, 'losses':0, 'draws':0}
 
   def process_output(line):
     # Parse line like this:
@@ -50,16 +50,17 @@ def run_games(run_id, run_chunk):
     if 'Score' in line:
       chunks = line.split(':')
       chunks = chunks[1].split()
-      state = {
-        'wins': int(chunks[0]),
-        'losses': int(chunks[2]),
-        'draws': int(chunks[4]),
-      }
+      state['wins'] = int(chunks[0])
+      state['losses'] = int(chunks[2])
+      state['draws'] = int(chunks[4])
 
       rundb.update_run_results(run_id, run_chunk, **state)
 
   # Run cutechess
-  sh.Command('./timed.sh')(run['worker_results'][run_chunk]['chunk_size'], run['args']['tc'], _out=process_output).wait()
+  chunk_size = run['worker_results'][run_chunk]['chunk_size']
+  p = sh.Command('./timed.sh')(chunk_size, run['args']['tc'], _out=process_output)
+  if p.exit_code() != 0:
+    raise Exception(p.stderr())
 
   if use_temp_dir:
     sh.rm('-rf', working_dir)
